requirements.txt
datasets==3.6.0
numpy==2.2.6
opencv_python_headless==4.11.0.86
peft==0.15.2
Pillow==11.2.1
qwen_vl_utils==0.0.11
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
